+++ validators/fmi/__init__.py
@@ -1,1 +1,0 @@
-# FMI validator package marker.