+++ validators/fmu/__init__.py
@@ -1,0 +1,1 @@
+# FMU validator package marker.